requirements.txt
# Minimum Requirements for a plotly dash app
Brotli==1.0.9
click==8.0.3
dash==2.15.0
dash-core-components==2.0.2
dash-html-components==2.0.2
plotly==5.3.1
Flask==2.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
plotly==5.3.1
six==1.15.0
tenacity==8.0.1
Werkzeug==1.0.1


# Used for my app
pandas==1.2.4
scipy==1.6.2
sentence_transformers==2.3.1
umap-learn==0.5.5
sklearn==0.0

# For Heroku
gunicorn==20.1.0